refactor(dashboard-auth): move home view to module
