+++ apps/dashboard/src/modules/home/views/HomeView.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script setup lang="ts">
-import BalanceWithTopupComponent from '@/components/BalanceWithTopupComponent.vue';
+import BalanceWithTopupComponent from '@/modules/home/components/BalanceWithTopup.vue';
 import MutationsTableComponent from '@/components/Mutations/MutationsTableComponent.vue';
 import { useAuthStore, useUserStore } from '@sudosos/sudosos-frontend-common';
 import apiService from '@/services/ApiService';
